Screen filed used on MacPro - portrait mode
I didn't have to tweak this on Ubuntu when I installed there.  After
struggling with it for a while, I managed to delete the file.  When I
launched the app, it created a screen.xml automatically for the
combined screens.  It must have pulled the info from the OS display
settings.  Anyway, I'm happy that it worked.